Fixed ViewResultBaseAssertions
ViewResultBase doesn't exist anymore in the AspNetCore.Mvc framework. This meant that a huge number of modification were needed to make the ViewResultBaseAssertions class work correctly.

Instead of polluting the whole class with tons of #if and to rename it to ViewResultAssertions, I conditionally removed the whole ViewResultBaseAssertions class and created a ViewResultAssertions only for the .Net Core project.