Made most of the unit test work
ViewResult objects are surprisingly complicated to create from scracth, especially with the ViewData collection. It's nearly impossible to create the collection "by hand".

Instead, I decided to try to simply create a TestController class that would return ViewResults and PartialViewResults constructed like I need for the tests. It makes passing the model and the viewdata much more simple.